Move Polyfill classes to separate directory and exclude them from Qodana
