+++ Src/FluentAssertions/Polyfill/StringBuilderExtensions.cs
@@ -1,8 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿#if NET47 || NETSTANDARD2_0 || NETSTANDARD2_1
 
-namespace FluentAssertions;
+using System.Collections.Generic;
+
+// ReSharper disable once CheckNamespace
+namespace System.Text;
 
 /// <summary>
 /// Since net6.0 StringBuilder has additional overloads taking an AppendInterpolatedStringHandler
@@ -15,8 +16,8 @@ internal static class StringBuilderExtensions
     public static StringBuilder AppendLine(this StringBuilder stringBuilder, IFormatProvider _, string value) =>
         stringBuilder.AppendLine(value);
 
-#if NET47 || NETSTANDARD2_0
     public static StringBuilder AppendJoin<T>(this StringBuilder stringBuilder, string separator, IEnumerable<T> values) =>
         stringBuilder.Append(string.Join(separator, values));
-#endif
 }
+
+#endif
